Relocated Liquid Global Context to its own directory
