+++ src/Umbraco.Community.Contentment.Client/src/global-context/index.ts
@@ -1,5 +1,5 @@
-// SPDX-License-Identifier: MPL-2.0
+// SPDX-License-Identifier: MIT
 // Copyright © 2025 Lee Kelleher
 
-export { CONTENTMENT_LIQUID_CONTEXT } from './liquid.context-token.js';
-export type { ContentmentLiquidContext } from './liquid.context.js';
+export { CONTENTMENT_LIQUID_CONTEXT } from './liquid/liquid.context-token.js';
+export type * from './liquid/liquid.context.js';
