+++ src/Umbraco.Community.Contentment.Client/src/global-context/manifests.ts
@@ -1,11 +1,6 @@
-// SPDX-License-Identifier: MPL-2.0
+// SPDX-License-Identifier: MIT
 // Copyright © 2025 Lee Kelleher
 
-export const manifests: Array<UmbExtensionManifest> = [
-	{
-		type: 'globalContext',
-		alias: 'Contentment.GlobalContext.Liquid',
-		name: '[Contentment] Liquid Context',
-		api: () => import('./liquid.context.js'),
-	},
-];
+import { manifests as liquid } from './liquid/manifests.js';
+
+export const manifests = [...liquid];
