+++ src/Umbraco.Community.Contentment.Client/src/property-editor-ui/templated-list/templated-list.element.ts
@@ -1,7 +1,7 @@
-// SPDX-License-Identifier: MPL-2.0
+// SPDX-License-Identifier: MIT
 // Copyright © 2024 Lee Kelleher
 
-import { CONTENTMENT_LIQUID_CONTEXT } from '../../global-context/index.js';
+import { CONTENTMENT_LIQUID_CONTEXT } from '../../global-context/liquid/liquid.context-token.js';
 import type { ContentmentListItem, ContentmentDataListOption } from '../types.js';
 import {
 	classMap,
